+++ assembly/blake3/index_simd.ts
@@ -1,5 +1,5 @@
 import { hashSimdScratch } from "./hasher_simd";
 
-export function hash(inPtr: usize, inLen: usize, outPtr: usize): void {
+export function hashUnsafe(inPtr: usize, inLen: usize, outPtr: usize): void {
   hashSimdScratch(inPtr, inLen, outPtr);
 }