+++ index.ts
@@ -1,0 +1,3 @@
+import * as blake3 from "./assembly/blake3";
+
+export { blake3 };